remove pacman workshop
the code in this workshop is very similar to the code here: http://www.grantjenks.com/docs/freegames/pacman.html. this means we are currently violating the apache license of that code. removing it from the workshops page while i contact the person who submitted it and get more context