feat: add search utility and integrate it with the frontend
add(./lib/search.js)

add search file that can filter data from the workshop data

```javascript

export const filter = curry((originalData, setData, searchString) => {
  /*****

   filter function code

  ******/
  setData(isEmpty(data) ? fallBack : data)
})
```

the above funtion filters data on the basis of search query provided by
user

Any of the following attribute having the search query parameter in it makes it
qulified to pass through the filter:

1. Section Title
2. Section Description
3. Workshop Title
4. Workshop Description
5. Workshop Search Tags

workshop seach tags are yaml attributes that a workshop can have which
can makes it appear on search query

you can think of them as seo tags but for the workshop search layer

add(./lib/search.js)

add **fallBack workshop data** to be shown when no workshop matches search
query

```javascript

const fallBack = [
  {
    title: 'No Workshops found ):',
    description:
      "Hey! We couldn't find a workshop for the provided search query. We update our workshops regularly and are our goal is to add workshops on as many topics as possible, so stay tuned :)",
    slugs: [],
    workshops: []
  }
]
```

the following is fallback data to be shown when the user query doesn't
match any workshop or section

add(./pages/index.js)

add styles to #__.next element to make it's min-height 100%

it fixes the *floating upwards of the footer* when there are no workshops
to show and fallback section has been rendered

```javascript
      <style jsx global>{`
        #__next {
          height: 100%;
        }
      `}</style>
```
the above code fixes that

how implementation of seach works?

the search function is curried and acts as a filteration layer

it is passed via props to our searchbox input element and is made to
update workshop data

the whole search layer works on the frontend and is the key to it being
so fast

all the changes in the codewase are backward compatible and support for
all the previous workshops is maintained.

Author: Harsh Bajpai
"May the code be with you :)"